rename var class
